YTDB-916 Move predicate-eval benches into core
Relocate the auxiliary predicate-evaluation microbenchmarks (Bench 1
evaluator sensitivity, Bench 2 FilterStep throughput, Bench 3 ExpandStep
throughput) plus their JMH state, the shared BenchDataset builder, and the
two JUnit correctness guards out of the jmh-ldbc LDBC standard-suite module
and into core test sources, alongside the existing core JMH microbenchmarks.

These benches are LDBC-free: they spin up an in-memory synthetic Bench
schema/star-graph and only exercise core query internals (AnalyzedExpr IR
lowering/evaluation, FilterStep/ExpandStep push-down). They never belonged in
the LDBC SNB module, whose jmh-ldbc placement forced awkward cross-module
indirection. Co-locating them in core lets the guard tests run as ordinary
core unit tests and reach the package-private ExpandStep.getAnalyzed() /
SubQueryStep.subExecutionPlan directly.

All 13 files move into the single package
com.jetbrains.youtrackdb.internal.core.sql.executor, which keeps BenchDataset
package-private (no visibility changes) and satisfies the guard test's
package-private access. The former ExpandStepIrProbe (a same-package helper
that only existed to cross the old module boundary) is now redundant, so its
plan-navigation logic is inlined into ExpandStepIrPathGuardTest and the probe
class is deleted. Redundant same-package imports and stale jmh-ldbc Javadoc
cross-references are cleaned up, and jmh-ldbc/README.md now points at the new
location instead of documenting these benches as living in that module.

No pom changes are required: core already wires the JMH BenchmarkProcessor
into test-compile and carries the jmh-core/jmh-generator-annprocess test deps.
