+++ core/src/test/java/com/jetbrains/youtrackdb/internal/core/sql/executor/AnalyzedExprEvaluatorBenchmark.java
@@ -1,4 +1,4 @@
-package com.jetbrains.youtrackdb.benchmarks.ldbc;
+package com.jetbrains.youtrackdb.internal.core.sql.executor;
 
 import com.jetbrains.youtrackdb.internal.core.query.Result;
 import com.jetbrains.youtrackdb.internal.core.query.analyzed.AnalyzedExprEvaluator;
