+++ core/src/test/java/com/jetbrains/youtrackdb/internal/core/sql/executor/AnalyzedExprEvaluatorSingleThreadBenchmark.java
@@ -1,4 +1,4 @@
-package com.jetbrains.youtrackdb.benchmarks.ldbc;
+package com.jetbrains.youtrackdb.internal.core.sql.executor;
 
 import org.openjdk.jmh.annotations.Threads;
 import org.openjdk.jmh.runner.Runner;
@@ -9,11 +9,9 @@ import org.openjdk.jmh.runner.options.OptionsBuilder;
  * Single-threaded runner for the Bench-1 predicate-evaluation sensitivity instrument
  * ({@link AnalyzedExprEvaluatorBenchmark}).
  *
- * <p>Usage:
- * <pre>
- * ./mvnw -pl jmh-ldbc -am verify -P bench -DskipTests \
- *   -Djmh.args="AnalyzedExprEvaluatorSingleThreadBenchmark.*"
- * </pre>
+ * <p>Now lives in {@code core} test sources; run it via this class's {@link #main(String[])} (JMH
+ * {@code Runner}/{@code OptionsBuilder}) — e.g. from an IDE or a test-classpath launcher — rather
+ * than through the old {@code jmh-ldbc} {@code -P bench} uber-jar.
  */
 @Threads(1)
 public class AnalyzedExprEvaluatorSingleThreadBenchmark extends AnalyzedExprEvaluatorBenchmark {
